Test use of plugins filename
