+++ test/closet/plugins/chairo.js
@@ -1,0 +1,6 @@
+'use strict';
+
+// plugins specified, but not an array and no register
+module.exports = {
+    plugins: { options: { web: false } }
+};
