+++ test/closet/plugins/my-dogwater.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const Dogwater = require('dogwater');
+
+module.exports = {
+    plugins: {
+        register: Dogwater,
+        options: {
+            adapters: { myAdapter: {} },
+            connections: { simple: { adapter: 'myAdapter' } }
+        }
+    }
+};
