+++ test/closet/plugins/my-loveboat.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const Loveboat = require('loveboat');
+
+// plugins specified as an array
+module.exports = {
+    plugins: [Loveboat]
+};
