+++ test/closet/plugins/vision.js
@@ -1,0 +1,4 @@
+'use strict';
+
+// No plugins specified
+module.exports = {};
